au -> ao and more
I changed "au" to "ao", changed stress to always be on the penultimate syllable when the last syllable is open (dunía, not dúnia), simplified the phonotactics, replaced my separate words for "this" and "that" ("ni" and "go") with separate words for "here" and "there" ("hir" and "der") (the generic demonstrative is now "da"), and added accute accents to a bunch of country names.  I also moved some mis-sorted terms and fixed the fact that a lot of words had -ta even tho I'm pretty sure I already changed that to -te.