fix causative verbs
I altered the description of the causative and passive suffixes. I realized the solution to my trubble was to simply have the adposition of the original subject of a causative verb change depending on whether the original verb was transitive or not.

I also started trying out this new notation where I use backticks to denote PUDUL text.  the idea is that if I ever make this into an actual wobsite, I'll rite a script to replace them all with links to their respective diccionary entries.  idk if the syntax will be the same, tho, since I would also probably be using typst in that senario.

I also droppd the suffix -ce/-co, because I realized I was using it for both "vehicle" and "wheel" which is hecka confusing.  also, just in general, vehicle names don't fully describe a vehicle (planes move themselves but aren't automobiles, cars have motors and wheels but aren't motorcycles, segways have two wheels but aren't bicycles, and of course other languages have equally nonsensical compound words), so I'm making a general shift from compounds plus -ce to roots that describe the vehicles inherently.   I added the names of some more peoples and religius structures, I moved the letter names to the category "zi nam" and put the actual letters in the category "zi" (this is mostly to form valid links when I describe the alphabet on the wobsite), I changed "base" to "nica", I changed "sude" to "sur", I aded more basic rooms and bildings (kind of embarassing I had forgotten to have a word for 'pharmasy'), anh I changed the pronoun pluralizer to mon.  I added words for different kinds of fee.  I changed "sana" to "meik" and made it the causative suffix.